ANP-85 Move API logic from HistoryViewModel to repository and implement Koin dependency injection for view model
