+++ android/app/src/main/java/hr/foi/air/wattsup/di/ViewModelsModule.kt
@@ -2,6 +2,7 @@ package hr.foi.air.wattsup.di
 
 import hr.foi.air.wattsup.viewmodels.AuthenticationViewModel
 import hr.foi.air.wattsup.viewmodels.CardViewModel
+import hr.foi.air.wattsup.viewmodels.HistoryViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
@@ -13,6 +14,11 @@ val viewModelsModule = module {
     }
     viewModel {
         CardViewModel(
+            repository = get(),
+        )
+    }
+    viewModel {
+        HistoryViewModel(
             repository = get(),
         )
     }
