+++ android/app/src/main/java/hr/foi/air/wattsup/repository/WattsUpRepositoryImpl.kt
@@ -5,6 +5,8 @@ import android.util.Log
 import hr.foi.air.wattsup.network.NetworkService
 import hr.foi.air.wattsup.network.models.Card
 import hr.foi.air.wattsup.network.models.CardPOSTBody
+import hr.foi.air.wattsup.network.models.Charger
+import hr.foi.air.wattsup.network.models.Event
 import hr.foi.air.wattsup.network.models.LoginBody
 import hr.foi.air.wattsup.network.models.LoginResponseBody
 import hr.foi.air.wattsup.network.models.RegistrationBody
@@ -13,6 +15,8 @@ import hr.foi.air.wattsup.network.models.TokenManager
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
+import kotlin.coroutines.resume
+import kotlin.coroutines.suspendCoroutine
 
 class WattsUpRepositoryImpl : WattsUpRepository {
     private val authService = NetworkService.authService
@@ -159,5 +163,91 @@ class WattsUpRepositoryImpl : WattsUpRepository {
                 }
             },
         )
+    }
+
+    override suspend fun getChargerName(context: Context, chargerId: Int): String? {
+        val chargerService = NetworkService.chargerService
+        val auth = "Bearer " + TokenManager.getInstance(context).getJWTToken()
+
+        return suspendCoroutine { continuation ->
+            chargerService.getCharger(chargerId, auth)
+                .enqueue(object : Callback<Charger?> {
+                    override fun onResponse(call: Call<Charger?>, response: Response<Charger?>) {
+                        if (response.isSuccessful) {
+                            continuation.resume(response.body()!!.name)
+                            Log.d("HistoryScreen", "Cards: ${response.body()}")
+                        } else {
+                            Log.d("HistoryScreen", "Error: ${response.errorBody()}")
+                            continuation.resume(null)
+                        }
+                    }
+
+                    override fun onFailure(call: Call<Charger?>, t: Throwable) {
+                        Log.d("HistoryScreen", "Failure: ${t.message}")
+                        continuation.resume(null)
+                    }
+                })
+        }
+    }
+
+    override suspend fun getCards(context: Context, userId: Int): List<Card?> {
+        val cardService = NetworkService.cardService
+        val auth = "Bearer " + TokenManager.getInstance(context).getJWTToken()
+
+        return suspendCoroutine { continuation ->
+            cardService.getCardsForUser(userId, auth)
+                .enqueue(object : Callback<List<Card?>> {
+                    override fun onResponse(
+                        call: Call<List<Card?>>,
+                        response: Response<List<Card?>>,
+                    ) {
+                        if (response.isSuccessful) {
+                            continuation.resume(response.body() ?: emptyList())
+                            Log.d("HistoryScreen", "Cards: ${response.body()}")
+                        } else {
+                            Log.d("HistoryScreen", "Error: ${response.errorBody()}")
+                            continuation.resume(emptyList())
+                        }
+                    }
+
+                    override fun onFailure(call: Call<List<Card?>>, t: Throwable) {
+                        Log.d("HistoryScreen", "Failure: ${t.message}")
+                        continuation.resume(emptyList())
+                    }
+                })
+        }
+    }
+
+    override suspend fun getEvents(
+        context: Context,
+        cardId: Int,
+        onResponse: (String) -> Unit,
+    ): List<Event?> {
+        val eventService = NetworkService.eventService
+        val auth = "Bearer " + TokenManager.getInstance(context).getJWTToken()
+
+        return suspendCoroutine { continuation ->
+            eventService.getEvents(cardId, auth).enqueue(object : Callback<List<Event?>> {
+                override fun onResponse(
+                    call: Call<List<Event?>>,
+                    response: Response<List<Event?>>,
+                ) {
+                    if (response.isSuccessful) {
+                        continuation.resume(response.body() ?: emptyList())
+                        Log.d("HistoryScreen", "Events: ${response.body()}")
+                    } else {
+                        Log.d("HistoryScreen", "Error: ${response.errorBody()}")
+                        onResponse("Error: ${response.errorBody()}")
+                        continuation.resume(emptyList())
+                    }
+                }
+
+                override fun onFailure(call: Call<List<Event?>>, t: Throwable) {
+                    Log.d("HistoryScreen", "Failure: ${t.message}")
+                    onResponse("Failure: ${t.message}")
+                    continuation.resume(emptyList())
+                }
+            })
+        }
     }
 }
